Various Config changes
Config file: 4 articles show on homepage.

ActiveStatus.php: allows the admin to set a message to users when
server is experiencing difficulty working properly. Also good for SQL
DIE statements and what not - good error handling.

editArticle.php - NEW LAYOUT BUT POSTING TO DB ENGINE IS BROKEN! PLEASE
USE OLD COPY OF THIS ONE FILE FOR NOW IF YOU NEED POSTING CAPABILITY,
WE ARE WORKING ON A FIX!